+++ Tests/Errors.test.ps1
@@ -34,8 +34,10 @@ task Warnings {
 	($r = Invoke-Build t1 $file -Result Result)
 
 	# output
-	assert ($r[-3] -clike 'WARNING: demo-file-warning File: *Errors.test.ps1.')
-	assert ($r[-2] -clike 'WARNING: demo-task-warning Task: t1. File: *Errors.test.ps1.')
+	assert ($r[-5] -cmatch '^WARNING: .*Errors\.test\.ps1:\d+$')
+	equals $r[-4] demo-file-warning
+	assert ($r[-3] -cmatch '^WARNING: /t1 .*Errors\.test\.ps1:\d+$')
+	equals $r[-2] demo-task-warning
 	assert ($r[-1] -clike 'Build succeeded with warnings. 1 tasks, 0 errors, 2 warnings *')
 
 	# result